Fixed Organization Chart color per rank. Added Contact List support using User Details linked to Custom Fields. Grouping by Department missing. User rights missing. User details edition missing.

+++ app/controllers/positions_controller.rb
@@ -5,8 +5,10 @@ class PositionsController < ApplicationController
 
   helper :hr
 
+  helper :custom_fields
+
   def index
-    @positions = Position.find(:all)
+    @positions = Position.find(:all,:order => 'position')
   end
 
 
@@ -19,22 +21,22 @@ class PositionsController < ApplicationController
   def new
     if request.get?
       @position = Position.new()
+      if params[:manager_id]
+        @position.manager_id=params[:manager_id]
+      end
     else
       @position = Position.new(params[:position])
       @position.title = params[:position][:title]
       @position.manager_id = params[:position][:manager_id]
       if @position.save
         (params[:user_id] || []).each { |user_id|
-        #        news.each { |new|
-        #          @role.workflows.build(:tracker_id => @tracker.id, :old_status_id => old, :new_status_id => new)
-        #        }
-        user_position=UserPosition.new
-        user_position.position_id=@position.id
-        user_position.user_id=user_id
-        user_position.save
+          @position.user_positions.build(:position_id => @position.id, :user_id => user_id)
         }
-        flash[:notice] = l(:notice_successful_create)
-        redirect_to(params[:continue] ? {:controller => 'positions', :action => 'new'} :
+
+        if @position.save
+          flash[:notice] = l(:notice_successful_update)
+        end
+        redirect_to(params[:continue] ? {:controller => 'positions', :action => 'new', :manager_id => @position.id} :
             {:controller => 'positions', :action => 'index', :id => @position})
         return
       end
@@ -44,23 +46,18 @@ class PositionsController < ApplicationController
   def edit
 
     @position = Position.find(params[:id])
-    if request.post?
-      @position.title = params[:position][:title] if params[:position][:title]
-      @position.manager_id = params[:position][:manager_id] if params[:position][:manager_id]
+    if request.post? and @position.update_attributes(params[:position])
+      
       UserPosition.destroy_all( ["position_id=?", @position.id])
+
       (params[:user_id] || []).each { |user_id|
-        #        news.each { |new|
-        #          @role.workflows.build(:tracker_id => @tracker.id, :old_status_id => old, :new_status_id => new)
-        #        }
-        user_position=UserPosition.new
-        user_position.position_id=@position.id
-        user_position.user_id=user_id
-        user_position.save
+        @position.user_positions.build(:position_id => @position.id, :user_id => user_id)
       }
+
       if @position.save
         flash[:notice] = l(:notice_successful_update)
-        redirect_to :action => 'index'
       end
+      redirect_to :action => 'index'
 
     end
   rescue ::ActionController::RedirectBackError
@@ -80,6 +77,9 @@ class PositionsController < ApplicationController
     @user_positions =  UserPosition.find(:all)
   end
 
+  def contact_list
+    @user_positions =  UserPosition.find(:all,:order => "position_id")
+  end
 
   
 end
